backend/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
docker==6.1.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.5